+++ apps/desktop/src/views/Welcome.test.tsx
@@ -18,6 +18,7 @@ function statusWith(over: Partial<Status> = {}): Status {
     steam_root: 'C:\\Program Files (x86)\\Steam',
     steam_running: true,
     offer_to_start_steam: true,
+    auto_start_steam: false,
     steam_source: 'HKCU',
     account_id: 1,
     has_api_key: false,